+++ NEWS.md
@@ -1,3 +1,9 @@
+# posDemux 0.99.6
+
+* Changed the data loader of the default callbacks, allowing for
+multiple input FASTQ files
+* Fixed bug where empty stringsets were not handled correctly
+
 # posDemux 0.99.4
 
 * First package version submitted to Bioconductor